Fix root mapping involving reductions
A reduction domain must not be mapped with any domain of any consumer
tensors as the loop for the reduction domain must be closed before any
of the consumers.

This restriction is applied by UnmappableReductionDomains. Previously,
it is only enforced when a domain is mappable with a reduction domain.
In that case, that domain is never made mappable with any of the
consumer domains. This is not however sufficient. We need to look at
all the domains used to generate the reduction domain. If a domain is
included in such a domain set, it must not be mappable with any consumer
domain. The domains that are used as input to a reduction domain is
found by FindInputDomains.